Updating Question Model and Routes
